+++ src/queries/store.ts
@@ -1,31 +1,12 @@
 import {
-  ApolloAction,
-  isQueryInitAction,
-  isQueryResultAction,
-  isQueryErrorAction,
-  isQueryResultClientAction,
-  isQueryStopAction,
-  isStoreResetAction,
-  StoreResetAction,
-} from '../actions';
-
-import {
-  graphQLResultHasError,
-} from '../data/storeUtils';
-
-import {
   DocumentNode,
-  SelectionSetNode,
   GraphQLError,
+  ExecutionResult,
 } from 'graphql';
 
 import { isEqual } from '../util/isEqual';
 
 import { NetworkStatus } from './networkStatus';
-
-export interface QueryStore {
-  [queryId: string]: QueryStoreValue;
-}
 
 export type QueryStoreValue = {
   queryString: string;
@@ -38,22 +19,31 @@ export type QueryStoreValue = {
   metadata: any;
 };
 
-export interface SelectionSetWithRoot {
-  id: string;
-  typeName: string;
-  selectionSet: SelectionSetNode;
-}
+export class QueryStore {
+  private store: {[queryId: string]: QueryStoreValue} = {};
 
-export function queries(
-  previousState: QueryStore = {},
-  action: ApolloAction,
-): QueryStore {
-  if (isQueryInitAction(action)) {
-    const newState = { ...previousState } as QueryStore;
+  public getStore(): {[queryId: string]: QueryStoreValue} {
+    return this.store;
+  }
 
-    const previousQuery = previousState[action.queryId];
+  public get(queryId: string): QueryStoreValue {
+    return this.store[queryId];
+  }
 
-    if (previousQuery && previousQuery.queryString !== action.queryString) {
+  public initQuery(query: {
+                     queryId: string,
+                     queryString: string,
+                     document: DocumentNode,
+                     storePreviousVariables: boolean,
+                     variables: Object,
+                     isPoll: boolean,
+                     isRefetch: boolean,
+                     metadata: any,
+                     fetchMoreForQueryId: string | undefined,
+                  }) {
+    const previousQuery = this.store[query.queryId];
+
+    if (previousQuery && previousQuery.queryString !== query.queryString) {
       // XXX we're throwing an error here to catch bugs where a query gets overwritten by a new one.
       // we should implement a separate action for refetching so that QUERY_INIT may never overwrite
       // an existing query (see also: https://github.com/apollostack/apollo-client/issues/732)
@@ -64,43 +54,42 @@ export function queries(
 
     let previousVariables: Object | null = null;
     if (
-      action.storePreviousVariables &&
+      query.storePreviousVariables &&
       previousQuery &&
       previousQuery.networkStatus !== NetworkStatus.loading
       // if the previous query was still loading, we don't want to remember it at all.
     ) {
-      if (!isEqual(previousQuery.variables, action.variables)) {
+      if (!isEqual(previousQuery.variables, query.variables)) {
         isSetVariables = true;
         previousVariables = previousQuery.variables;
       }
     }
 
     // TODO break this out into a separate function
-    let newNetworkStatus = NetworkStatus.loading;
-
+    let networkStatus;
     if (isSetVariables) {
-      newNetworkStatus = NetworkStatus.setVariables;
-    } else if (action.isPoll) {
-      newNetworkStatus = NetworkStatus.poll;
-    } else if (action.isRefetch) {
-      newNetworkStatus = NetworkStatus.refetch;
+      networkStatus = NetworkStatus.setVariables;
+    } else if (query.isPoll) {
+      networkStatus = NetworkStatus.poll;
+    } else if (query.isRefetch) {
+      networkStatus = NetworkStatus.refetch;
       // TODO: can we determine setVariables here if it's a refetch and the variables have changed?
-    } else if (action.isPoll) {
-      newNetworkStatus = NetworkStatus.poll;
+    } else {
+      networkStatus = NetworkStatus.loading;
     }
 
     // XXX right now if QUERY_INIT is fired twice, like in a refetch situation, we just overwrite
     // the store. We probably want a refetch action instead, because I suspect that if you refetch
     // before the initial fetch is done, you'll get an error.
-    newState[action.queryId] = {
-      queryString: action.queryString,
-      document: action.document,
-      variables: action.variables,
+    this.store[query.queryId] = {
+      queryString: query.queryString,
+      document: query.document,
+      variables: query.variables,
       previousVariables,
       networkError: null,
       graphQLErrors: [],
-      networkStatus: newNetworkStatus,
-      metadata: action.metadata,
+      networkStatus,
+      metadata: query.metadata,
     };
 
     // If the action had a `moreForQueryId` property then we need to set the
@@ -110,121 +99,71 @@ export function queries(
     // error action branch, but importantly *not* in the client result branch.
     // This is because the implementation of `fetchMore` *always* sets
     // `fetchPolicy` to `network-only` so we would never have a client result.
-    if (typeof action.fetchMoreForQueryId === 'string') {
-      newState[action.fetchMoreForQueryId] = {
-        // We assume that that both a query with id `action.moreForQueryId`
-        // already exists and that it is not `action.queryId`. This is a safe
-        // assumption given how we set `moreForQueryId`.
-        ...previousState[action.fetchMoreForQueryId],
-        // We set the network status to `fetchMore` here overwriting any
-        // network status that currently exists. This is how network statuses
-        // are set normally, so it makes sense to set it this way here as well.
-        networkStatus: NetworkStatus.fetchMore,
-      };
+    if (typeof query.fetchMoreForQueryId === 'string') {
+      this.store[query.fetchMoreForQueryId].networkStatus = NetworkStatus.fetchMore;
     }
-
-    return newState;
-  } else if (isQueryResultAction(action)) {
-    if (!previousState[action.queryId]) {
-      return previousState;
-    }
-
-    const newState = { ...previousState } as QueryStore;
-    const resultHasGraphQLErrors = graphQLResultHasError(action.result);
-
-    newState[action.queryId] = {
-      ...previousState[action.queryId],
-      networkError: null,
-      graphQLErrors: resultHasGraphQLErrors ? action.result.errors : [],
-      previousVariables: null,
-      networkStatus: NetworkStatus.ready,
-    };
-
-    // If we have a `fetchMoreForQueryId` then we need to update the network
-    // status for that query. See the branch for query initialization for more
-    // explanation about this process.
-    if (typeof action.fetchMoreForQueryId === 'string') {
-      newState[action.fetchMoreForQueryId] = {
-        ...previousState[action.fetchMoreForQueryId],
-        networkStatus: NetworkStatus.ready,
-      };
-    }
-
-    return newState;
-  } else if (isQueryErrorAction(action)) {
-    if (!previousState[action.queryId]) {
-      return previousState;
-    }
-
-    const newState = { ...previousState } as QueryStore;
-
-    newState[action.queryId] = {
-      ...previousState[action.queryId],
-      networkError: action.error,
-      networkStatus: NetworkStatus.error,
-    };
-
-    // If we have a `fetchMoreForQueryId` then we need to update the network
-    // status for that query. See the branch for query initialization for more
-    // explanation about this process.
-    if (typeof action.fetchMoreForQueryId === 'string') {
-      newState[action.fetchMoreForQueryId] = {
-        ...previousState[action.fetchMoreForQueryId],
-        networkError: action.error,
-        networkStatus: NetworkStatus.error,
-      };
-    }
-
-    return newState;
-  } else if (isQueryResultClientAction(action)) {
-    if (!previousState[action.queryId]) {
-      return previousState;
-    }
-
-    const newState = { ...previousState } as QueryStore;
-
-    newState[action.queryId] = {
-      ...previousState[action.queryId],
-      networkError: null,
-      previousVariables: null,
-      // XXX I'm not sure what exactly action.complete really means. I assume it means we have the complete result
-      // and do not need to hit the server. Not sure when we'd fire this action if the result is not complete, so that bears explanation.
-      // We should write that down somewhere.
-      networkStatus: action.complete ? NetworkStatus.ready : NetworkStatus.loading,
-    };
-
-    return newState;
-  } else if (isQueryStopAction(action)) {
-    const newState = { ...previousState } as QueryStore;
-
-    delete newState[action.queryId];
-    return newState;
-  } else if (isStoreResetAction(action)) {
-    return resetQueryState(previousState, action);
   }
 
-  return previousState;
-}
+  public markQueryResult(queryId: string, result: ExecutionResult, fetchMoreForQueryId: string | undefined) {
+    if (!this.store[queryId]) {
+      return;
+    }
 
-// Returns the new query state after we receive a store reset action.
-// Note that we don't remove the query state for the query IDs that are associated with watchQuery()
-// observables. This is because these observables are simply refetched and not
-// errored in the event of a store reset.
-function resetQueryState(state: QueryStore, action: StoreResetAction): QueryStore {
-  const observableQueryIds = action.observableQueryIds;
+    this.store[queryId].networkError = null;
+    this.store[queryId].graphQLErrors = (result.errors && result.errors.length) ? result.errors : [];
+    this.store[queryId].previousVariables = null;
+    this.store[queryId].networkStatus = NetworkStatus.ready;
 
-  // keep only the queries with query ids that are associated with observables
-  const newQueries = Object.keys(state).filter((queryId) => {
-    return (observableQueryIds.indexOf(queryId) > -1);
-  }).reduce((res, key) => {
-    // XXX set loading to true so listeners don't trigger unless they want results with partial data
-    res[key] = {
-      ...state[key],
-      networkStatus: NetworkStatus.loading,
-    };
+    // If we have a `fetchMoreForQueryId` then we need to update the network
+    // status for that query. See the branch for query initialization for more
+    // explanation about this process.
+    if (typeof fetchMoreForQueryId === 'string') {
+      this.store[fetchMoreForQueryId].networkStatus = NetworkStatus.ready;
+    }
+  }
 
-    return res;
-  }, {} as QueryStore);
+  public markQueryError(queryId: string, error: Error, fetchMoreForQueryId: string | undefined) {
+    if (!this.store[queryId]) {
+      return;
+    }
 
-  return newQueries;
+    this.store[queryId].networkError = error;
+    this.store[queryId].networkStatus = NetworkStatus.error;
+
+    // If we have a `fetchMoreForQueryId` then we need to update the network
+    // status for that query. See the branch for query initialization for more
+    // explanation about this process.
+    if (typeof fetchMoreForQueryId === 'string') {
+      this.markQueryError(fetchMoreForQueryId, error, undefined);
+    }
+  }
+
+  public markQueryResultClient(queryId: string, complete: boolean) {
+    if (!this.store[queryId]) {
+      return;
+    }
+
+    this.store[queryId].networkError = null;
+    this.store[queryId].previousVariables = null;
+    this.store[queryId].networkStatus = complete ? NetworkStatus.ready : NetworkStatus.loading;
+  }
+
+  public stopQuery(queryId: string) {
+    delete this.store[queryId];
+  }
+
+  public reset(observableQueryIds: string[]) {
+    // keep only the queries with query ids that are associated with observables
+    this.store = Object.keys(this.store).filter((queryId) => {
+      return (observableQueryIds.indexOf(queryId) > -1);
+    }).reduce((res, key) => {
+      // XXX set loading to true so listeners don't trigger unless they want results with partial data
+      res[key] = {
+        ...this.store[key],
+        networkStatus: NetworkStatus.loading,
+      };
+
+      return res;
+    }, {} as {[queryId: string]: QueryStoreValue});
+  }
 }