+++ src/store.ts
@@ -20,12 +20,11 @@ import {
 } from './data/storeUtils';
 
 import {
-  queries,
   QueryStore,
 } from './queries/store';
 
 import {
-  mutations,
+  // mutations,
   MutationStore,
 } from './mutations/store';
 
@@ -62,8 +61,6 @@ export interface ReducerError {
 
 export interface Store {
   data: NormalizedCache;
-  queries: QueryStore;
-  mutations: MutationStore;
   optimistic: OptimisticStore;
   reducerError: ReducerError | null;
 }
@@ -111,9 +108,6 @@ export function createApolloReducer(config: ApolloReducerConfig): (state: Store,
   return function apolloReducer(state = {} as Store, action: ApolloAction) {
     try {
       const newState: Store = {
-        queries: queries(state.queries, action),
-        mutations: mutations(state.mutations, action),
-
         data: data(state.data, action, config),
         optimistic: [] as any,
 
@@ -136,8 +130,6 @@ export function createApolloReducer(config: ApolloReducerConfig): (state: Store,
       );
 
       if (state.data === newState.data &&
-      state.mutations === newState.mutations &&
-      state.queries === newState.queries &&
       state.optimistic === newState.optimistic &&
       state.reducerError === newState.reducerError) {
         return state;