+++ test/client.ts
@@ -181,8 +181,6 @@ describe('client', () => {
       client.store.getState(),
       {
         apollo: {
-          queries: {},
-          mutations: {},
           data: {},
           optimistic: [],
           reducerError: null,
@@ -576,19 +574,6 @@ describe('client', () => {
     };
 
     const finalState = { apollo: assign({}, initialState.apollo, {
-      queries: {
-        '1': {
-          queryString: print(query),
-          document: query,
-          variables: {},
-          networkStatus: NetworkStatus.ready,
-          networkError: null,
-          graphQLErrors: [],
-          previousVariables: null,
-          metadata: null,
-        },
-      },
-      mutations: {},
       reducerError: null,
     }) };
 
@@ -1533,6 +1518,66 @@ describe('client', () => {
     // if deduplication didn't happen, result.data will equal data2.
     return Promise.all([q1, q2]).then(([result1, result2]) => {
       assert.deepEqual(result1.data, result2.data);
+    });
+  });
+
+  it('emits Redux actions when the flag is enabled', () => {
+    QueryManager.EMIT_REDUX_ACTIONS = true;
+
+    const query = gql`
+      query people {
+        allPeople(first: 1) {
+          people {
+            name
+            __typename
+          }
+          __typename
+        }
+      }
+    `;
+
+    const data = {
+      allPeople: {
+        people: [
+          {
+            name: 'Luke Skywalker',
+            __typename: 'Person',
+          },
+        ],
+        __typename: 'People',
+      },
+    };
+
+    const networkInterface = mockNetworkInterface({
+      request: { query: cloneDeep(query) },
+      result: { data },
+    });
+
+    const client = new ApolloClient({
+      networkInterface,
+    });
+
+    client.initStore();
+
+    const orig = client.store.dispatch;
+    let actionEmitted = false;
+
+    client.store.dispatch = (action) => {
+      if (action.type === 'APOLLO_QUERY_INIT') {
+        actionEmitted = true;
+      }
+
+      orig(action);
+    };
+
+    const queryPromise = client.query({ query }).then((result) => {
+      assert.deepEqual(result.data, data);
+    });
+
+    QueryManager.EMIT_REDUX_ACTIONS = false;
+
+    return queryPromise.then(() => {
+      assert(actionEmitted, 'An action was not emitted');
     });
   });